requirements.txt
streamlit>=1.28.0
openai>=1.6.1
faiss-cpu>=1.7.0
langchain>=0.0.350
langchain-openai>=0.0.2
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.24.0
tiktoken>=0.5.0
beautifulsoup4>=4.12.0 